Organize db access files
